requirements.txt
cosmotech-api>=2.0.0
azure-identity==1.12.0
pyyaml=*
azure-cosmos=*